Round 3 of the great rangesel cursor cleanup Still some bugs but things are much closer
2001-05-08  Jody Goldberg <jgoldberg@home.com>

	* src/selection.c (sheet_selection_set_internal) : split this out into.
	* src/sheet-merge.c (sheet_merge_find_container) : here.
	* src/sheet-control-gui.c (scg_rangesel_changed) : and here.

2001-05-08  Jody Goldberg <jgoldberg@home.com>

	* src/xml-io.c (xml_read_selection_info) : Use sheet_set_edit_pos.

	* src/selection.c (sheet_selection_extend) : merge into
	  scg_cursor_extend.

	* src/sheet-control-gui.c (scg_cursor_move_to) : rename from
	  scg_cursor_move.
	(scg_cursor_extend) : merge scg_cursor_extend_[hv]
	(scg_cursor_move) : merge scg_cursor_move_[hv]
	(scg_rangesel_extend) : merge scg_rangesel_extend_[hv]
	(scg_rangesel_move) : merge scg_rangesel_move_[hv]
	(scg_rangesel_start) : cleanup.

	* src/gnumeric-sheet.c (gnumeric_sheet_key_mode_sheet) : fix handling
	  of Home. Simplify the function pointers for move vs extend by
	  merging horizontal and vertical versions.

+++ plugins/numtheory/numtheory.c
@@ -21,7 +21,7 @@
 GNUMERIC_MODULE_PLUGIN_INFO_DECL;
 
 #ifndef WORD_BIT
-#define WORD_BIT (sizeof (int) * CHAR_BIT)
+#define WORD_BIT ((int)(sizeof (int) * CHAR_BIT))
 #endif
 
 #define OUT_OF_BOUNDS "#LIMIT!"
